Fix Glow & Outline Shader on AMD

+++ 1.8.9-Forge/src/main/resources/assets/minecraft/liquidbounce/shader/fragment/outline.frag
@@ -4,7 +4,7 @@ uniform sampler2D texture;
 uniform vec2 texelSize;
 
 uniform vec4 color;
-uniform int radius;
+uniform float radius;
 
 void main(void) {
     vec4 centerCol = texture2D(texture, gl_TexCoord[0].xy);
@@ -12,8 +12,8 @@ void main(void) {
     if(centerCol.a != 0) {
         gl_FragColor = vec4(0, 0, 0, 0);
     } else {
-        for (int x = -radius; x <= radius; x++) {
-            for (int y = -radius; y <= radius; y++) {
+        for (float x = -radius; x <= radius; x++) {
+            for (float y = -radius; y <= radius; y++) {
                 vec4 currentColor = texture2D(texture, gl_TexCoord[0].xy + vec2(texelSize.x * x, texelSize.y * y));
 
                 if (currentColor.a != 0) {
